+++ modules/layer_norm.cpp
@@ -1,7 +1,7 @@
 #include <hls_math.h>
 
 template<int M, int N>
-void layer_norm(double out[M][N], double in[M][N], double eps) {
+void layer_norm(float out[M][N], float in[M][N], float eps) {
     #pragma HLS INTERFACE port=return mode=s_axilite
     #pragma HLS INTERFACE port=out mode=bram
     #pragma HLS INTERFACE port=in mode=bram
@@ -10,8 +10,8 @@ void layer_norm(double out[M][N], double in[M][N], double eps) {
     for (int i = 0; i < M; i++) {
         #pragma HLS PIPELINE II=1
 
-        double mean = 0.0;
-        double variance = 0.0;
+        float mean = 0.0;
+        float variance = 0.0;
 
         // Step 1: 計算均值
         for (int j = 0; j < N; j++) {
@@ -23,12 +23,12 @@ void layer_norm(double out[M][N], double in[M][N], double eps) {
         // Step 2: 計算變異數
         for (int j = 0; j < N; j++) {
             #pragma HLS UNROLL
-            double diff = in[i][j] - mean;
+            float diff = in[i][j] - mean;
             variance += diff * diff;
         }
         variance /= N;
 
-        double stddev_inv = 1.0 / hls::sqrt(variance + eps);
+        float stddev_inv = 1.0 / hls::sqrt(variance + eps);
 
         // Step 3: 標準化
         for (int j = 0; j < N; j++) {
@@ -39,24 +39,24 @@ void layer_norm(double out[M][N], double in[M][N], double eps) {
 }
 
 template<int M, int N>
-void rms_norm(double out[N][M], double in[N][M], double gamma[M]) {
+void rms_norm(float out[N][M], float in[N][M], float gamma[M]) {
     #pragma HLS INTERFACE port=return mode=s_axilite
     #pragma HLS INTERFACE port=in mode=bram
     #pragma HLS INTERFACE port=out mode=bram
     #pragma HLS INTERFACE port=gamma mode=bram
 
-    const double epsilon = 1e-5;
+    const float epsilon = 1e-5;
 
     for (int n = 0; n < N; n++) {
         #pragma HLS PIPELINE II=1
 
-        double sq_sum = 0.0;
+        float sq_sum = 0.0;
         for (int i = 0; i < M; i++) {
             #pragma HLS UNROLL
             sq_sum += in[n][i] * in[n][i];
         }
 
-        double rms = hls::sqrt(sq_sum / M + epsilon);
+        float rms = hls::sqrt(sq_sum / M + epsilon);
 
         for (int i = 0; i < M; i++) {
             #pragma HLS UNROLL
