+++ modules/transformer.hpp
@@ -11,7 +11,7 @@ template <int N, class T>
 void maxval_1D(T* max_val, T in[N]);
 
 template <int M, int N>
-void softmax(double out[M][N], double in[M][N]);
+void softmax(float out[M][N], float in[M][N]);
 
 template <int M, int N, class T>
 void transpose(T out[N][M], T in[M][N]);
@@ -23,7 +23,7 @@ template <int M, int K, int N, class T>
 void qkv_generator(T out_q[M][N], T out_k[M][N], T out_v[M][N], T in_x[M][K], T in_wq[K][N], T in_wk[K][N], T in_wv[K][N]);
 
 template<int M, int N>
-void layer_norm(double out[M][N], double in[M][N]);
+void layer_norm(float out[M][N], float in[M][N]);
 
 template <int M, int N, class T>
 void feed_forward(T out[M][N], T in_x[M][N], T in_w1[N][4 * N], T in_w2[4 * N][N], T in_b1[M][4 * N], T in_b2[M][N]);
@@ -32,7 +32,7 @@ template <int M, int N, class T>
 void relu(T out[M][N], T in[M][N]);
 
 template <int M, int N, class T>
-void elementwise_mul(T out[M][N], T in[M][N], double in_scaling);
+void elementwise_mul(T out[M][N], T in[M][N], float in_scaling);
 
 template <int in_features, int out_features, int n_samples, class T>
 void linear(T out[n_samples][out_features], T in[n_samples][in_features], T weight[out_features][in_features], T bias[out_features]);
